# src/index.ts
import cluster from 'node:cluster';
import express, { Request, Response, NextFunction } from 'express';
import rateLimit from "express-rate-limit";
import helmet from "helmet";
import { pinoHttp } from "pino-http";
import crypto from "crypto";
import { LRUCache } from "lru-cache";
import { z } from 'zod';
import swaggerUi from 'swagger-ui-express';
import { readFileSync } from 'node:fs';
import { join } from 'node:path';

// Load Swagger JSON
const swaggerDocument = JSON.parse(
    readFileSync(join(process.cwd(), 'src', 'swagger.json'), 'utf8')
);

const app = express();
const PORT = process.env.PORT || 8080;

// Validation Schemas
const weatherParamsSchema = z.object({
    zip: z.string().regex(/^\d{5}$/, "Invalid format. Zip code must be exactly 5 digits.")
});
// Defines the temperature scale constraint
const weatherQuerySchema = z.object({
    scale: z.preprocess(
        (val) => (val === "" ? "Fahrenheit" : val),
        // Ensures the scale is a string and defaults to Fahrenheit if empty
        z.string().optional().default("Fahrenheit")
    ).refine(
        (val) => ["celsius", "fahrenheit"].includes((val as string).toLowerCase()),
        { message: "Invalid scale. Must be Celsius or Fahrenheit" }
    ) as z.ZodType<string>
});

//Helmet middleware for security headers
app.use(helmet());

//Limits each IP address to 100 requests per 15-minute window 
const limiter = rateLimit({
    windowMs: 15 * 60 * 1000,
    max: 100,
    standardHeaders: 'draft-8',
    legacyHeaders: false,
    message: { error: 'Too many requests, please try again later.' },
});
app.use(limiter)

// Avoid noisy request logging during automated tests.
if (process.env.NODE_ENV !== "test") {
    app.use(pinoHttp({
        genReqId: (req) => req.headers['x-request-id'] || crypto.randomUUID(),
        customAttributeKeys: {
            reqId: 'requestId'
        }
    }));
}

// Correlation ID Middleware (Ensures headers are set even if pino is off)
app.use((req: Request, res: Response, next: NextFunction) => {
    const requestId = (req.id as string) || (req.headers['x-request-id'] as string) || crypto.randomUUID();
    req.headers['x-request-id'] = requestId;
    res.setHeader('X-Request-ID', requestId);
    next();
});


//LRU Cache to store the last 100 requests for 10 minutes
const cache = new LRUCache<string, { temperature: number; scale: string }>({
    max: 100,
    ttl: 10 * 60 * 1000, //10 minutes
})

// Health check endpoint
app.get('/health', (req, res) => {
    res.status(200).json({
        status: 'UP',
        uptime: process.uptime(),
        timestamp: new Date().toISOString()
    });
});

// Swagger Documentation
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));


app.get('/locations/:zip', async (req: Request, res: Response, next: NextFunction) => {
    try {
        // Validate Parameters and Query using Zod
        const paramsResult = weatherParamsSchema.safeParse(req.params);
        const queryResult = weatherQuerySchema.safeParse(req.query);

        if (!paramsResult.success) {
            return res.status(400).json({ error: paramsResult.error.issues[0]?.message || "Invalid Parameters" });
        }
        if (!queryResult.success) {
            return res.status(400).json({ error: queryResult.error.issues[0]?.message || "Invalid Query" });
        }

        const { zip: zipCode } = paramsResult.data;
        const normalizedScale = queryResult.data.scale.toLowerCase();

        // Check cache for existing result
        const cacheKey = `${zipCode}-${normalizedScale}`;
        const cached = cache.get(cacheKey);
        if (cached) {
            //If cache hit, return the stored result without calling any API
            res.status(200).json(cached);
            return;
        }


        //Build the GeoCoding API URL using string interpolation
        const geoUrl = `https://geocoding-api.open-meteo.com/v1/search?name=${zipCode}&count=1&language=en&format=json`;

        //Fetch the data and wait for the network to respond 
        const geoResponse = await fetch(geoUrl);

        //Pass the raw text response into a structured JSON onject
        const geoData = await geoResponse.json();

        //Protection against user typing an invalid zip code and the API finding nothing
        if (!geoData.results || geoData.results.length == 0) {
            res.status(404).json({ error: "Location not found for the provided zip code." });
            return
        }
        //Extract the exact coordinates from the JSON array
        const lat = geoData.results[0].latitude;
        const lon = geoData.results[0].longitude;
        const apiUnit = normalizedScale === 'celsius' ? 'celsius' : 'fahrenheit';
        const finalScale = normalizedScale === 'celsius' ? 'Celsius' : 'Fahrenheit';
        const weatherUrl = `https://api.open-meteo.com/v1/forecast?latitude=${lat}&longitude=${lon}&current_weather=true&temperature_unit=${apiUnit}`;
        const weatherResponse = await fetch(weatherUrl);
        const weatherData = await weatherResponse.json() as any;
        if (!weatherData || !weatherData.current_weather) {
            return res.status(502).json({ error: "Weather API returned no data" });
        }
        const result = {
            temperature: Math.round(weatherData.current_weather.temperature),
            scale: finalScale

        }
        //Store in cache for future request
        cache.set(cacheKey, result);

        //Send the final successful response back to the user
        res.status(200).json(result);


    }

    //If the network drops, return a clean error

    catch (error) {
        next(error);
    }

});

// 404 Handler for undefined routes
app.use((req: Request, res: Response, next: NextFunction) => {
    const error: any = new Error('Resource Not Found');
    error.status = 404;
    next(error);
});

// Centralized Error Handling Middleware
app.use((err: any, req: Request, res: Response, next: NextFunction) => {
    // Handle Zod Validation Errors
    if (err instanceof z.ZodError) {
        return res.status(400).json({
            error: err.issues[0]?.message || "Validation Error",
            status: 400,
            timestamp: new Date().toISOString()
        });
    }

    const status = err.status || 500;
    const message = err.message || 'Internal Server Error';

    // Log the error using pino-http logger if available
    if (req.log) {
        req.log.error(err);
    } else {
        console.error("Critical Error:", err);
    }

    res.status(status).json({
        error: message,
        status,
        timestamp: new Date().toISOString()
    });
});
// Only start the server if this  file is run directly( not by Jest)
if (process.env.NODE_ENV != 'test') {
    app.listen(PORT, () => {
        if (!cluster.isWorker) {
            console.log(`TypeScript Server is successfully running on http://localhost:${PORT}`);
        }
    });
}
export default app;